Update DPFClient to be non-throwing and delegate to a provider
